+++ lib/constant/contants.dart
@@ -1,8 +1,11 @@
 // extension sdk version
-const String SbExtensionKeyUIKit = 'sb_uikit';
-const String SbExtensionKeySyncManager = 'sb_syncmanager';
+const String sbExtensionKeyUIKit = 'sb_uikit';
+const String sbExtensionKeySyncManager = 'sb_syncmanager';
 
-const String SbExtraDataPremiumFeatureList = 'premium_feature_list';
-const String SbExtraDataFileUploadSizeLimit = 'file_upload_size_limit';
-const String SbExtraDataEmojiHash = 'emoji_hash';
-const String SbExtraDataApplicationAttributes = 'application_attributes';
+const String sbExtraDataPremiumFeatureList = 'premium_feature_list';
+const String sbExtraDataFileUploadSizeLimit = 'file_upload_size_limit';
+const String sbExtraDataEmojiHash = 'emoji_hash';
+const String sbExtraDataApplicationAttributes = 'application_attributes';
+
+const String sbPushTemplateDefault = 'default';
+const String sbPushTemplateAlternative = 'alternative';