+++ lib/query/application_user_list_query.dart
@@ -3,17 +3,25 @@ part of 'user_list_query.dart';
 /// A query object to retrieve list of user on current application.
 @JsonSerializable()
 class ApplicationUserListQuery extends UserListQuery {
-  /// The meta data key filter. This query will return users
-  /// that has the meta data key and values
+  /// Filters users who have this meta data key.
+  ///
+  /// This query will return users that has the meta data key and values
+  /// This should be specified in conjunction with the metaDataValues.
   String metaDataKey;
 
-  /// The meta data values filter. This query will return users
-  /// that has the meta data key and values
+  /// Filters users who have this meta data values.
+  ///
+  /// This query will return users that has the meta data key and values
+  /// This should be specified in conjunction with the metaDataKey.
   List<String> metaDataValues;
+
+  /// Filters users whose nicknames start with the this value
+  String nicknameStartsWith;
 
   ApplicationUserListQuery({
     this.metaDataKey,
     this.metaDataValues,
+    this.nicknameStartsWith,
   }) : super(queryType: UserListQueryType.filtered);
 
   @override
@@ -29,6 +37,7 @@ class ApplicationUserListQuery extends UserListQuery {
       userIds: userIds,
       metaDataKey: metaDataKey,
       metaDataValues: metaDataValues,
+      nicknameStartsWith: nicknameStartsWith,
     );
 
     loading = false;